refactor(ReservationRepository): ReservationRepository 추상화
